Loop is not derived from NodeInterface??
