--- src/ast/node/control/control.hpp
@@ -1,15 +1,0 @@
-#ifndef CROW_AST_NODE_CONTROL_CONTROL_HPP
-#define CROW_AST_NODE_CONTROL_CONTROL_HPP
-
-
-namespace ast::node::control {
-class If;
-class Loop;
-
-class Continue;
-class Break;
-class Defer;
-class Return;
-}; // namespace node::control
-
-#endif // CROW_AST_NODE_CONTROL_CONTROL_HPP
