+++ src/ast/node/control/defer.hpp
@@ -5,7 +5,7 @@
 #include "../node_traits/include.hpp"
 
 // Local Includes:
-#include "control.hpp"
+#include "fdecl.hpp"
 
 
 namespace ast::node::control {
