+++ src/ast/node/control/fdecl.hpp
@@ -1,0 +1,15 @@
+#ifndef FDECL_HPP
+#define FDECL_HPP
+
+
+namespace ast::node::control {
+class If;
+class Loop;
+
+class Continue;
+class Break;
+class Defer;
+class Return;
+}; // namespace ast::node::control
+
+#endif // FDECL_HPP
