+++ src/ast/node/control/if.hpp
@@ -5,7 +5,7 @@
 #include "../node_interface.hpp"
 
 // Local Includes:
-#include "control.hpp"
+#include "fdecl.hpp"
 
 
 namespace ast::node::control {
