+++ src/ast/node/control/include.hpp
@@ -1,0 +1,12 @@
+#ifndef INCLUDE_HPP
+#define INCLUDE_HPP
+
+// Control Statements:
+#include "break.hpp"
+#include "continue.hpp"
+#include "defer.hpp"
+#include "if.hpp"
+#include "loop.hpp"
+#include "return.hpp"
+
+#endif // INCLUDE_HPP
