+++ src/ast/node/control/loop.hpp
@@ -5,7 +5,7 @@
 #include "../node_traits/include.hpp"
 
 // Local Includes:
-#include "control.hpp"
+#include "fdecl.hpp"
 
 
 namespace ast::node::control {
@@ -13,9 +13,8 @@ namespace ast::node::control {
 namespace nt = node_traits;
 
 // Classes:
-class Loop : public nt::Body {
+class Loop : public nt::Init, public nt::Body {
   private:
-  NodePtr m_init;
   NodePtr m_condition;
   NodePtr m_expr;
 
@@ -23,7 +22,6 @@ class Loop : public nt::Body {
   Loop(NodePtr&& t_init, NodePtr&& t_condition, NodePtr&& t_expr,
        NodeListPtr&& t_body);
 
-  auto init() -> NodePtr&;
   auto condition() -> NodePtr&;
   auto expr() -> NodePtr&;
