+++ src/ast/node/node_traits/CMakeLists.txt
@@ -1,5 +1,6 @@
 # Add source files
 target_sources(${PROJECT_NAME} PRIVATE
+	init.cpp
 	identifier.cpp
 	body.cpp
 )
