+++ src/ast/node/node_traits/include.hpp
@@ -2,6 +2,7 @@
 #define CROW_AST_NODE_NODE_TRAITS_INCLUDE_HPP
 
 // Node traits:
+#include "init.hpp"
 #include "body.hpp"
 #include "identifier.hpp"
