+++ src/ast/node/typing/fdecl.hpp
@@ -1,0 +1,11 @@
+#ifndef FDECL_HPP
+#define FDECL_HPP
+
+
+namespace ast::node::typing {
+class Struct;
+class Interface;
+class Impl;
+} // namespace ast::node::typing
+
+#endif // FDECL_HPP
