+++ src/ast/node/typing/impl.hpp
@@ -4,6 +4,9 @@
 // Includes:
 #include "../node_traits/include.hpp"
 
+// Local Includes:
+#include "fdecl.hpp"
+
 
 namespace ast::node::typing {
 // Aliases:
