+++ src/ast/node/typing/include.hpp
@@ -1,0 +1,10 @@
+#ifndef INCLUDE_HPP
+#define INCLUDE_HPP
+
+// Typing:#
+#include "impl.hpp"
+#include "interface.hpp"
+#include "struct.hpp"
+#include "dot_expr.hpp"
+
+#endif // INCLUDE_HPP
