--- src/ast/node/typing/typing.hpp
@@ -1,11 +1,0 @@
-#ifndef CROW_AST_NODE_TYPING_TYPING_HPP
-#define CROW_AST_NODE_TYPING_TYPING_HPP
-
-
-namespace ast::node::typing {
-class Struct;
-class Interface;
-class Impl;
-} // namespace ast::node::typing
-
-#endif // CROW_AST_NODE_TYPING_TYPING_HPP
